+++ etl/socios/cultura/agentes_libro/dimensions_pipeline.py
@@ -31,7 +31,7 @@ class DimAgentesPipeline(EasyPipeline):
         db_connector = Connector.fetch('clickhouse-database', open('../../conns.yaml'))
 
         dtype = {
-            params.get('pk'): 'UInt8'
+            params.get('pk'): 'UInt16'
         }
 
         transform_step = TransformStep()
